+++ src/orderbook/asset_swapper_orderbook.ts
@@ -1,5 +1,6 @@
 import { FillQuoteTransformerOrderType, Orderbook, SignedLimitOrder } from '../asset-swapper';
 import { DEFAULT_PAGE, DEFAULT_PER_PAGE } from '../constants';
+import { logger } from '../logger';
 import { OrderBookService } from '../services/orderbook_service';
 import { SRAOrder } from '../types';
 
@@ -8,25 +9,46 @@ export class AssetSwapperOrderbook extends Orderbook {
         super();
     }
 
+    /**
+     * Returns matching limit orders.
+     *
+     *  NOTE: If the request to orderbook fails, it returns an empty array.
+     */
     public async getOrdersAsync(
         makerToken: string,
         takerToken: string,
         pruneFn?: (o: SignedLimitOrder) => boolean,
     ): Promise<SignedLimitOrder[]> {
-        const apiOrders = await this.orderbookService.getOrdersAsync(
-            DEFAULT_PAGE,
-            DEFAULT_PER_PAGE,
-            {
-                makerToken,
-                takerToken,
-            },
-            {},
-        );
+        const apiOrders = await this.orderbookService
+            .getOrdersAsync(
+                DEFAULT_PAGE,
+                DEFAULT_PER_PAGE,
+                {
+                    makerToken,
+                    takerToken,
+                },
+                {},
+            )
+            .catch((err) => {
+                logger.warn(
+                    {
+                        takerToken,
+                        makerToken,
+                        errorMessage: err.message,
+                    },
+                    'Request to Orderbook failed',
+                );
+
+                return {
+                    records: [],
+                };
+            });
 
         const orders = apiOrders.records.map(apiOrderToOrderbookOrder);
         const result = pruneFn ? orders.filter(pruneFn) : orders;
         return result;
     }
+
     public async getBatchOrdersAsync(
         makerTokens: string[],
         takerToken: string,
@@ -42,6 +64,7 @@ export class AssetSwapperOrderbook extends Orderbook {
         );
         return groupedByMakerToken;
     }
+
     public async destroyAsync(): Promise<void> {
         return;
     }
