+++ src/runners/utils.ts
@@ -28,8 +28,8 @@ import { OrderBookService } from '../services/orderbook_service';
 import { PostgresRfqtFirmQuoteValidator } from '../services/postgres_rfqt_firm_quote_validator';
 import { SwapService } from '../services/swap_service';
 import { HttpServiceConfig, AppDependencies } from '../types';
-import { AssetSwapperOrderbook } from '../utils/asset_swapper_orderbook';
-import { NoOpOrderbook } from '../utils/no_op_orderbook';
+import { AssetSwapperOrderbook } from '../orderbook/asset_swapper_orderbook';
+import { NoOpOrderbook } from '../orderbook/no_op_orderbook';
 import { RfqClient } from '../utils/rfq_client';
 import { RfqDynamicBlacklist } from '../utils/rfq_dyanmic_blacklist';
 import { S3Client } from '../utils/s3_client';
